Issue #22 - added libopendnp3-java Debian package support
The C++ library that provides the JNI support is now built against the
Oracle Java 7 runtime and packaged into libopendnp3-java.

Note that on many distributions (e.g. Ubuntu), Oracle Java 7 is not used
by default so must be added explicitly.  For anyone who needs to install
the official Oracle Java 7 runtime on a Debian system, see the
following:

http://webupd8.org/2012/01/install-oracle-java-jdk-7-in-ubuntu-via.html