Replace vcxproj with CMake project files
CMake can generate project files for (almost) any compiler toolkit you might
be interested in - and in this case, for the one I use as well as for the one
Adam appears to be using.

Boost is found by CMake - easier if you set the BOOST_ROOT env. var.

WFM but YMMV

Signed-off-by: Ronald Landheer-Cieslak <rlc@vlinder.ca>