Remove Database ownership from Outstation
The Outstation class, context, etc. are now passed a pointer to a
IDatabase, which is a now-decoupled interface for the database
implemented by the Database class -- or possibly something else.

The Outstation classes no longer know how to make a Databas, so they no
longer need the template, nor some of the database-specific
configuration parameters -- all of which have been removed as
appropriate.

This de-couples the database implementation from the Outstation
implementation.