Remove Modify method from MeasUpdate and Database
This method is not called anywhere in the code and is not the way
MeasUpdate is "supposed" to be used. It appears to be there only for a
specific corner-case for the dotnet bindings, which IMO should be handled
within the dotnet bindings in stead (there is no reason to pollute the
code with it for the benefit of .NET).

This also makes it easier to refactor MeasUpdate, as suggested here:
https://groups.google.com/d/msg/automatak-dnp3/f4kaYP7gFrw/fijxjMI_BQAJ
because there will simply be less code to work with.

It further makes it easier to integrate with a database that doesn't know
about ASIO and doesn't do the call-back thing.