Refactor MeasUpdate to MeasurementUpdate
This adds the commit() method.
Adding rollback(), if we ever want it, would be a simple matter of
	delete m_changes
but there are  no use-cases for this as yet.

All call sites have been changed. The class was renamed to make that
easier. The file was renamed to follow the class name, as appears to be
the rule for this project. (It's a good rule to have, IMHO).

This makes Meas[urement]Update exception-safe

It still allocates the m_changes member, which is a ChangeSet which in
turn contains an std::vector, which in turn contains instances of
std::function - so plenty of actual and potential allocations to go
around. Allocating m_changes is inevitable because it may survive the
MeasurementUpdate object (because the commit is asynchronous), although we
could wait for it or use some non-allocating queueing mechanism -- but we
can also just leave that for other database implementations to do...