A cleaner way to update flags
As noted by Adam here: http://bit.ly/1RX5iXH and here:
http://bit.ly/1RX5lCT, we need a way to update flags without updating the
value. This code ddoes that.

I've added a no_value_change_tag type that can be passed to the
constructor (conveniently, using no_value_change), which will have the
database ignore whatever value is encoded in the measurements and take
only the flags and the timestamp.