Change `CMAKE_SOURCE_DIR` to `PROJECT_SOURCE_DIR`.
I created a new project that will use OpenDNP3. After checking-out a Git
submodule under "vendor/dnp3", I added the following lines to my CMakeLists.txt:

    add_subdirectory(vendor/dnp3)
    include_directories(vendor/dnp3/cpp/libs/include)

This, however, led to an error:

    Error: include could not find load file:
        /Users/<user>/<project root>/cmake/settings.cmake

According to [this][1], `CMAKE_SOURCE_DIR` is always the project root,
which in this case is my project's root, not OpenDNP3's root. Changing to
`PROJECT_SOURCE_DIR` resolves this issue.

[1]: http://stackoverflow.com/questions/22214349/modify-cmake-source-dir