Fix for #396 LinkContext shutdown datarace
The problem is that the link keepalive and response timers on LinkContext
could have outstanding handlers at the time of shutdown.
Most of the time this is OK, because the handlers get cancelled on OnLowerLayerDown()
and usually execute with operation aborted error code.

But in the case where a timer handler is already queued for execution
when OnLowerLayerDown() is called, it executes with a success error code
and goes on to access LinkContext members which are in the process of being destroyed.

This fix copies a weak pointer to the ILinkSession into the handlers.
The link session abstraction is the top level of the composition that
the LinkContext is a part of, and hence lifetime is tied to.
This way the handlers can check if their logical lifetime has ended and return
instead of racing with the destruction of the LinkContext.